+++ frontend/src/_components/common/header-navigation/index.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useAuth } from '@/hooks/useAuth'
+import { useAuth } from '@/hooks/use-Auth'
 import { AppBar, Container, Link, List, ListItem } from '@mui/material'
 
 export const HeaderNavigation = () => {
