+++ frontend/src/app/page.tsx
@@ -4,7 +4,7 @@ import { HeaderNavigation } from '@/_components/common/header-navigation'
 
 import { TodoList } from '@/_components/features/todo/list'
 
-import { useAuth } from '@/hooks/useAuth'
+import { useAuth } from '@/hooks/use-Auth'
 import AppSuspense from '@/_components/common/app-suspense'
 import { TodoAddArea } from '@/_components/features/todo/add-area'
 import { Box, Container, Typography } from '@mui/material'
